+++ src/main/kotlin/gwanlang/Break.kt
@@ -1,0 +1,3 @@
+package gwanlang
+
+class Break : RuntimeException(null, null, true, false)